fix: move permissions file to the features/auth folder
