+++ app/features/user/schema.ts
@@ -1,8 +1,9 @@
 import { t } from 'i18next';
 import { z } from 'zod';
 
-import { zRole } from '@/lib/auth/permissions';
 import { zu } from '@/lib/zod/zod-utils';
+
+import { zRole } from '@/features/auth/permissions';
 
 export type User = z.infer<ReturnType<typeof zUser>>;
 export const zUser = () =>
