+++ app/lib/auth/client.ts
@@ -8,7 +8,7 @@ import { createAuthClient } from 'better-auth/react';
 import { envClient } from '@/env/client';
 import type { Auth } from '@/server/auth';
 
-import { permissions } from './permissions';
+import { permissions } from '../../features/auth/permissions';
 
 export const authClient = createAuthClient({
   baseURL: `${envClient.VITE_BASE_URL}/api/auth`,
@@ -20,7 +20,3 @@ export const authClient = createAuthClient({
     emailOTPClient(),
   ],
 });
-
-export type Permission = NonNullable<
-  Parameters<typeof authClient.admin.hasPermission>['0']['permission']
->;
